requirements.txt
# Database (scan results)
duckdb>=0.8.0

# Thumbnails (opencv + Pillow)
opencv-python>=4.5.0
Pillow>=9.0.0

# Charts (analysis/chart)
matplotlib>=3.5.0

# HTML template (Jinja2)
Jinja2>=3.0.0
